functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy
opencv-python
pyzbar
gevent
requests
certifi
charset-normalizer
idna
urllib3
pep8
autopep8
